feat: Add event filtering and provider interfaces, implement in-memory event store for testing
