+++ src/Sekiban.Dcb/Actors/EventProviderState.cs
@@ -1,0 +1,14 @@
+namespace Sekiban.Dcb.Actors;
+
+/// <summary>
+///     State of the event provider
+/// </summary>
+public enum EventProviderState
+{
+    NotStarted,
+    CatchingUp,
+    Live,
+    Paused,
+    Stopped,
+    Error
+}
