+++ src/Sekiban.Dcb/Actors/IEventSubscription.cs
@@ -47,74 +47,3 @@ public interface IEventSubscription : IDisposable
         string? subscriptionId = null,
         CancellationToken cancellationToken = default);
 }
-
-/// <summary>
-///     Handle for managing an event subscription
-/// </summary>
-public interface IEventSubscriptionHandle : IDisposable
-{
-    /// <summary>
-    ///     Unique ID for this subscription
-    /// </summary>
-    string SubscriptionId { get; }
-
-    /// <summary>
-    ///     Whether the subscription is currently active
-    /// </summary>
-    bool IsActive { get; }
-
-    /// <summary>
-    ///     Get the current position/checkpoint of the subscription
-    /// </summary>
-    string? CurrentPosition { get; }
-
-    /// <summary>
-    ///     Pause the subscription
-    /// </summary>
-    Task PauseAsync();
-
-    /// <summary>
-    ///     Resume the subscription
-    /// </summary>
-    Task ResumeAsync();
-
-    /// <summary>
-    ///     Unsubscribe and clean up resources
-    /// </summary>
-    Task UnsubscribeAsync();
-}
-
-/// <summary>
-///     Filter for events
-/// </summary>
-public interface IEventFilter
-{
-    /// <summary>
-    ///     Check if an event should be included
-    /// </summary>
-    bool ShouldInclude(Event evt);
-}
-
-/// <summary>
-///     Filter events by type
-/// </summary>
-public record EventTypeFilter(HashSet<string> EventTypes) : IEventFilter
-{
-    public bool ShouldInclude(Event evt) => EventTypes.Contains(evt.EventType);
-}
-
-/// <summary>
-///     Filter events by tags
-/// </summary>
-public record EventTagFilter(HashSet<string> RequiredTags) : IEventFilter
-{
-    public bool ShouldInclude(Event evt) => RequiredTags.All(tag => evt.Tags.Contains(tag));
-}
-
-/// <summary>
-///     Composite filter that combines multiple filters with AND logic
-/// </summary>
-public record CompositeEventFilter(List<IEventFilter> Filters) : IEventFilter
-{
-    public bool ShouldInclude(Event evt) => Filters.All(f => f.ShouldInclude(evt));
-}
